chill.requirements.txt
black==22.3.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.7.1
django-livereload-server==0.3.2
django-rest-framework==0.1.0
flake8==3.8.4
gunicorn==20.0.4
ipython==8.4.0
pip-chill==1.0.1
psycopg2-binary==2.8.6
pylint==2.6.0
python-decouple==3.4
termcolor==1.1.0
